+++ include/franka/robot_state.h
@@ -62,6 +62,21 @@ struct RobotState {
   std::array<double, 16> EE_T_K{};  // NOLINT (readability-identifier-naming)
 
   /**
+   * Configured mass of the end effector.
+   */
+  double m_ee{};
+
+  /**
+   * Configured rotational inertia matrix of the end effector load w.r.t.\ center of mass.
+   */
+  std::array<double, 9> I_ee{};  // NOLINT (readability-identifier-naming)
+
+  /**
+   * Configured center of mass of the end effector load w.r.t.\ flange frame.
+   */
+  std::array<double, 3> F_x_Cee{};  // NOLINT (readability-identifier-naming)
+
+  /**
    * Configured mass of the external load.
    */
   double m_load{};
@@ -75,6 +90,22 @@ struct RobotState {
    * Configured center of mass of the external load w.r.t.\ flange frame.
    */
   std::array<double, 3> F_x_Cload{};  // NOLINT (readability-identifier-naming)
+
+  /**
+   * Sum of the mass of the end effector and the external load.
+   */
+  double m_total{};
+
+  /**
+   * Combined rotational inertia matrix of the end effector load and the external load w.r.t.\
+   * center of mass.
+   */
+  std::array<double, 9> I_total{};  // NOLINT (readability-identifier-naming)
+
+  /**
+   * Combined center of mass of the end effector load and the external load w.r.t.\ flange frame.
+   */
+  std::array<double, 3> F_x_Ctotal{};  // NOLINT (readability-identifier-naming)
 
   /**
    * Elbow configuration.