+++ test/helpers.cpp
@@ -6,77 +6,93 @@
 
 #include <gtest/gtest.h>
 
+#include "load_calculations.h"
+
 bool stringContains(const std::string& actual, const std::string& expected) {
   return actual.find(expected) != std::string::npos;
 }
 
 void testRobotStateIsZero(const franka::RobotState& actual) {
-  for (size_t i = 0; i < actual.O_T_EE.size(); i++) {
-    EXPECT_EQ(0.0, actual.O_T_EE[i]);
+  for (double element : actual.O_T_EE) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.O_T_EE_d.size(); i++) {
-    EXPECT_EQ(0.0, actual.O_T_EE_d[i]);
+  for (double element : actual.O_T_EE_d) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.EE_T_K.size(); i++) {
-    EXPECT_EQ(0.0, actual.EE_T_K[i]);
+  for (double element : actual.EE_T_K) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.F_T_EE.size(); i++) {
-    EXPECT_EQ(0.0, actual.F_T_EE[i]);
+  for (double element : actual.F_T_EE) {
+    EXPECT_EQ(0.0, element);
+  }
+  EXPECT_EQ(0.0, actual.m_ee);
+  for (double element : actual.F_x_Cee) {
+    EXPECT_EQ(0.0, element);
+  }
+  for (double element : actual.I_ee) {
+    EXPECT_EQ(0.0, element);
   }
   EXPECT_EQ(0.0, actual.m_load);
-  for (size_t i = 0; i < actual.F_x_Cload.size(); i++) {
-    EXPECT_EQ(0.0, actual.F_x_Cload[i]);
+  for (double element : actual.F_x_Cload) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.I_load.size(); i++) {
-    EXPECT_EQ(0.0, actual.I_load[i]);
+  for (double element : actual.I_load) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.elbow.size(); i++) {
-    EXPECT_EQ(0.0, actual.elbow[i]);
+  EXPECT_EQ(0.0, actual.m_total);
+  for (double element : actual.F_x_Ctotal) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.elbow_d.size(); i++) {
-    EXPECT_EQ(0.0, actual.elbow_d[i]);
+  for (double element : actual.I_total) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.tau_J.size(); i++) {
-    EXPECT_EQ(0.0, actual.tau_J[i]);
+  for (double element : actual.elbow) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.tau_J_d.size(); i++) {
-    EXPECT_EQ(0.0, actual.tau_J_d[i]);
+  for (double element : actual.elbow_d) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.dtau_J.size(); i++) {
-    EXPECT_EQ(0.0, actual.dtau_J[i]);
+  for (double element : actual.tau_J) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.q.size(); i++) {
-    EXPECT_EQ(0.0, actual.q[i]);
+  for (double element : actual.tau_J_d) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.dq.size(); i++) {
-    EXPECT_EQ(0.0, actual.dq[i]);
+  for (double element : actual.dtau_J) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.q_d.size(); i++) {
-    EXPECT_EQ(0.0, actual.q_d[i]);
+  for (double element : actual.q) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.dq_d.size(); i++) {
-    EXPECT_EQ(0.0, actual.dq_d[i]);
+  for (double element : actual.dq) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.joint_contact.size(); i++) {
-    EXPECT_EQ(0.0, actual.joint_contact[i]);
+  for (double element : actual.q_d) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.cartesian_contact.size(); i++) {
-    EXPECT_EQ(0.0, actual.cartesian_contact[i]);
+  for (double element : actual.dq_d) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.joint_collision.size(); i++) {
-    EXPECT_EQ(0.0, actual.joint_collision[i]);
+  for (double element : actual.joint_contact) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.cartesian_collision.size(); i++) {
-    EXPECT_EQ(0.0, actual.cartesian_collision[i]);
+  for (double element : actual.cartesian_contact) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.tau_ext_hat_filtered.size(); i++) {
-    EXPECT_EQ(0.0, actual.tau_ext_hat_filtered[i]);
+  for (double element : actual.joint_collision) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.O_F_ext_hat_K.size(); i++) {
-    EXPECT_EQ(0.0, actual.O_F_ext_hat_K[i]);
+  for (double element : actual.cartesian_collision) {
+    EXPECT_EQ(0.0, element);
   }
-  for (size_t i = 0; i < actual.K_F_ext_hat_K.size(); i++) {
-    EXPECT_EQ(0.0, actual.K_F_ext_hat_K[i]);
+  for (double element : actual.tau_ext_hat_filtered) {
+    EXPECT_EQ(0.0, element);
+  }
+  for (double element : actual.O_F_ext_hat_K) {
+    EXPECT_EQ(0.0, element);
+  }
+  for (double element : actual.K_F_ext_hat_K) {
+    EXPECT_EQ(0.0, element);
   }
   EXPECT_FALSE(actual.current_errors);
   EXPECT_FALSE(actual.last_motion_errors);
@@ -89,9 +105,15 @@ void testRobotStatesAreEqual(const franka::RobotState& expected, const franka::R
   EXPECT_EQ(expected.O_T_EE_d, actual.O_T_EE_d);
   EXPECT_EQ(expected.F_T_EE, actual.F_T_EE);
   EXPECT_EQ(expected.EE_T_K, actual.EE_T_K);
+  EXPECT_EQ(expected.m_ee, actual.m_ee);
+  EXPECT_EQ(expected.F_x_Cee, actual.F_x_Cee);
+  EXPECT_EQ(expected.I_ee, actual.I_ee);
   EXPECT_EQ(expected.m_load, actual.m_load);
   EXPECT_EQ(expected.F_x_Cload, actual.F_x_Cload);
   EXPECT_EQ(expected.I_load, actual.I_load);
+  EXPECT_EQ(expected.m_total, actual.m_total);
+  EXPECT_EQ(expected.F_x_Ctotal, actual.F_x_Ctotal);
+  EXPECT_EQ(expected.I_total, actual.I_total);
   EXPECT_EQ(expected.elbow, actual.elbow);
   EXPECT_EQ(expected.elbow_d, actual.elbow_d);
   EXPECT_EQ(expected.tau_J, actual.tau_J);
@@ -121,9 +143,20 @@ void testRobotStatesAreEqual(const research_interface::robot::RobotState& expect
   EXPECT_EQ(expected.O_T_EE_d, actual.O_T_EE_d);
   EXPECT_EQ(expected.F_T_EE, actual.F_T_EE);
   EXPECT_EQ(expected.EE_T_K, actual.EE_T_K);
+  EXPECT_EQ(expected.m_ee, actual.m_ee);
+  EXPECT_EQ(expected.F_x_Cee, actual.F_x_Cee);
+  EXPECT_EQ(expected.I_ee, actual.I_ee);
   EXPECT_EQ(expected.m_load, actual.m_load);
   EXPECT_EQ(expected.F_x_Cload, actual.F_x_Cload);
   EXPECT_EQ(expected.I_load, actual.I_load);
+  EXPECT_EQ(expected.m_ee + expected.m_load, actual.m_total);
+  EXPECT_EQ(franka::combineCenterOfMass(expected.m_ee, expected.F_x_Cee, expected.m_load,
+                                        expected.F_x_Cload),
+            actual.F_x_Ctotal);
+  EXPECT_EQ(franka::combineInertiaTensor(expected.m_ee, expected.F_x_Cee, expected.I_ee,
+                                         expected.m_load, expected.F_x_Cload, expected.I_load,
+                                         actual.m_total, actual.F_x_Ctotal),
+            actual.I_total);
   EXPECT_EQ(expected.elbow, actual.elbow);
   EXPECT_EQ(expected.elbow_d, actual.elbow_d);
   EXPECT_EQ(expected.tau_J, actual.tau_J);
@@ -193,11 +226,25 @@ void randomRobotState(franka::RobotState& robot_state) {
   for (double& element : robot_state.F_T_EE) {
     element = randomDouble();
   }
+  robot_state.m_ee = randomDouble();
+  for (double& element : robot_state.F_x_Cee) {
+    element = randomDouble();
+  }
+  for (double& element : robot_state.I_ee) {
+    element = randomDouble();
+  }
   robot_state.m_load = randomDouble();
   for (double& element : robot_state.F_x_Cload) {
     element = randomDouble();
   }
   for (double& element : robot_state.I_load) {
+    element = randomDouble();
+  }
+  robot_state.m_total = randomDouble();
+  for (double& element : robot_state.F_x_Ctotal) {
+    element = randomDouble();
+  }
+  for (double& element : robot_state.I_total) {
     element = randomDouble();
   }
   for (double& element : robot_state.elbow) {
@@ -274,6 +321,13 @@ void randomRobotState(research_interface::robot::RobotState& robot_state) {
     element = randomDouble();
   }
   for (double& element : robot_state.F_T_EE) {
+    element = randomDouble();
+  }
+  robot_state.m_ee = randomDouble();
+  for (double& element : robot_state.F_x_Cee) {
+    element = randomDouble();
+  }
+  for (double& element : robot_state.I_ee) {
     element = randomDouble();
   }
   robot_state.m_load = randomDouble();